% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{avg_event_price_line_plot}
\alias{avg_event_price_line_plot}
\title{Plot Average Ticket Price Over Last 200 Events}
\usage{
avg_event_price_line_plot(events_df)
}
\arguments{
\item{events_df}{A data frame containing event data, including \code{Min_Price}, \code{Max_Price}, and \code{Event_Name} columns.}
}
\value{
An interactive ggplotly object displaying the average ticket price trend over the last 200 events.
}
\description{
Generates a line plot showing the average ticket price trend for the last 200 events.
}
\examples{
\dontrun{
avg_event_price_line_plot(events_df)
}
}
