% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{avg_price_class_plot}
\alias{avg_price_class_plot}
\title{Plot Average Ticket Price by Event Type for the Last 200 Events}
\usage{
avg_price_class_plot(events_df)
}
\arguments{
\item{events_df}{A data frame containing event data, including \code{Min_Price}, \code{Max_Price}, and \code{Genre} columns.}
}
\value{
A ggplot object showing the average ticket price for each event genre.
}
\description{
Generates a bar chart displaying the average ticket price for events, grouped by genre.
}
\examples{
\dontrun{
avg_price_class_plot(events_df)
}
}
