% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{event_class_plot}
\alias{event_class_plot}
\title{Plot Events by Genre}
\usage{
event_class_plot(events_df)
}
\arguments{
\item{events_df}{A data frame containing event data retrieved using \code{get_full_ticketmaster_data()}.}
}
\value{
A ggplot object displaying the number of events by genre.
}
\description{
Generates a bar chart showing the number of events available per genre.
}
\examples{
\dontrun{
event_class_plot(events_df)
}
}
