% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{get_ticketmaster_api_key}
\alias{get_ticketmaster_api_key}
\title{Retrieve Ticketmaster API Key}
\usage{
get_ticketmaster_api_key()
}
\value{
A character string containing the API key, or stops execution if not found.
}
\description{
Fetches the API key from environment variables.
}
\examples{
api_key <- get_ticketmaster_api_key()
print(api_key)
}
