% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_code.R
\name{ticketmaster_analysis}
\alias{ticketmaster_analysis}
\title{Ticketmaster Event Analysis}
\usage{
ticketmaster_analysis(events_df)
}
\arguments{
\item{events_df}{A data frame containing event data retrieved using \code{get_full_ticketmaster_data()}.}
}
\value{
A list containing:
\itemize{
\item \code{most_events_venue}: The venue with the most events.
\item \code{least_events_venue}: The venue with the least events.
\item \code{highest_ticket_price_event}: The event with the highest ticket price.
\item \code{highest_ticket_price}: The highest ticket price.
\item \code{lowest_ticket_price_event}: The event with the lowest ticket price.
\item \code{lowest_ticket_price}: The lowest ticket price.
}
}
\description{
Analyzes event data from Ticketmaster, including highest and lowest ticket prices and venue statistics.
}
\examples{
\dontrun{
result <- ticketmaster_analysis(events_df)
print(result)
}
}
